+++ simpletransformers/ner/ner_model.py
@@ -478,13 +478,15 @@ class NERModel:
         loss = outputs[0]
         if self.loss_fct:
             logits = outputs[1]
-            labels = inputs['labels']
-            attention_mask = inputs.get('attention_mask')
+            labels = inputs["labels"]
+            attention_mask = inputs.get("attention_mask")
             if attention_mask is not None:
                 active_loss = attention_mask.view(-1) == 1
                 active_logits = logits.view(-1, self.num_labels)
                 active_labels = torch.where(
-                    active_loss, labels.view(-1), torch.tensor(self.loss_fct.ignore_index).type_as(labels)
+                    active_loss,
+                    labels.view(-1),
+                    torch.tensor(self.loss_fct.ignore_index).type_as(labels),
                 )
                 loss = self.loss_fct(active_logits, active_labels)
             else: