+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.61.10] - 2021-07-13
+
+### Added
+
+- Added class weights support for NER [tiena2cva](https://github.com/tiena2cva)
+- Added Herbert [Zhylkaaa](https://github.com/Zhylkaaa)
+
+### Fixed
+
+- Bug fixes
+
 ## [0.61.9] - 2021-06-21
 
 ### Fixed